fix: added SceneDelegate to avoid futur crash with iOS 27
